demo/src/test/java/com/example/demo/OrderProductServiceImplTest.java

package com.example.demo;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.*;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import com.example.demo.models.OrderProduct;
import com.example.demo.repositories.OrderProductRepository;
import com.example.demo.repositories.OrderRepository;
import com.example.demo.services.impl.OrderProductServiceImpl;

@ExtendWith(MockitoExtension.class)
public class OrderProductServiceImplTest {

    @Mock
    private OrderProductRepository orderProductRepository;

    @Mock
    private OrderRepository orderRepository;

    @InjectMocks
    private OrderProductServiceImpl orderProductService;

    private OrderProduct orderProduct;

    @BeforeEach
    public void setup() {
        // Initialize test data
        orderProduct = new OrderProduct();
        orderProduct.setId(1L);
        // Set other properties as needed
    }

    @Test
    public void testCreateOrderProduct() {
        // Stubbing behavior for findAll method
        when(orderProductRepository.findAll()).thenReturn(new ArrayList<>());

        // Stubbing behavior for save method
        when(orderProductRepository.save(orderProduct)).thenReturn(orderProduct);

        // Test the createOrderProduct method
        OrderProduct createdOrderProduct = orderProductService.createOrderProduct(orderProduct);

        // Verify the result
        assertNotNull(createdOrderProduct);
        assertEquals(orderProduct, createdOrderProduct);
    }

    @Test
    public void testGetOrderProduct() {
        // Stubbing behavior for findById method
        when(orderProductRepository.findById(1L)).thenReturn(Optional.of(orderProduct));

        // Test the getOrderProduct method
        OrderProduct foundOrderProduct = orderProductService.getOrderProduct(1L);

        // Verify the result
        assertNotNull(foundOrderProduct);
        assertEquals(orderProduct, foundOrderProduct);
    }

    @Test
    public void testUpdateOrderProduct() {
        // Stubbing behavior for findById method
        when(orderProductRepository.findById(1L)).thenReturn(Optional.of(orderProduct));

        // Stubbing behavior for save method
        when(orderProductRepository.save(any())).thenReturn(orderProduct);

        // Test the updateOrderProduct method
        OrderProduct updatedOrderProduct = new OrderProduct();
        updatedOrderProduct.setId(1L);
        // Set other properties as needed

        OrderProduct result = orderProductService.updateOrderProduct(1L, updatedOrderProduct);

        // Verify the result
        assertNotNull(result);
        assertEquals(updatedOrderProduct.getId(), result.getId());
        // Add more assertions for other properties if needed
    }

    @Test
    public void testDeleteOrderProduct() {
        // Stubbing behavior for findById method
        when(orderProductRepository.findById(1L)).thenReturn(Optional.of(orderProduct));

        // Test the deleteOrderProduct method
        boolean result = orderProductService.deleteOrderProduct(1L);

        // Verify the result
        assertTrue(result);
        verify(orderProductRepository, times(1)).delete(orderProduct);
    }

    @Test
    public void testFindByOrderId() {
        // Stubbing behavior for findByOrderId method
        when(orderProductRepository.findByOrderId(1L)).thenReturn(new ArrayList<>());

        // Test the findByOrderId method
        List<OrderProduct> result = orderProductService.findByOrderId(1L);

        // Verify the result
        assertNotNull(result);
        assertTrue(result.isEmpty());
    }

    @Test
    public void testGetAllOrderProducts() {
        // Stubbing behavior for findAll method
        List<OrderProduct> orderProducts = new ArrayList<>();
        orderProducts.add(orderProduct);
        when(orderProductRepository.findAll()).thenReturn(orderProducts);

        // Test the getAllOrderProducts method
        List<OrderProduct> result = orderProductService.getAllOrderProducts();

        // Verify the result
        assertNotNull(result);
        assertFalse(result.isEmpty());
        assertEquals(1, result.size());
    }

    @Test
    public void testGetOrderProductsByOrderId() {
        // Stubbing behavior for findByOrderId method
        List<OrderProduct> orderProducts = new ArrayList<>();
        orderProducts.add(orderProduct);
        when(orderProductRepository.findByOrderId(1L)).thenReturn(orderProducts);

        // Test the getOrderProductsByOrderId method
        List<OrderProduct> result = orderProductService.getOrderProductsByOrderId(1L);

        // Verify the result
        assertNotNull(result);
        assertFalse(result.isEmpty());
        assertEquals(1, result.size());
    }

    // Write similar tests for updateOrderProduct, deleteOrderProduct, findByOrderId, getAllOrderProducts, getOrderProductsByOrderId
}
